@@ -2,23 +2,16 @@ package com.consultant.model.controller;
 
 import com.consultant.model.dto.UserDTO;
 import com.consultant.model.entities.User;
-import com.consultant.model.security.JwtResponse;
-import com.consultant.model.security.JwtTokenUtil;
-import com.consultant.model.security.JwtUserDetailsService;
-import com.consultant.model.security.WrongValidationException;
+import com.consultant.model.exception.EmailMissingException;
+import com.consultant.model.exception.ForbiddenException;
+import com.consultant.model.mappers.UserMapper;
 import com.consultant.model.services.impl.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.core.ResolvableType;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.oauth2.client.registration.ClientRegistration;
-import org.springframework.security.oauth2.client.registration.ClientRegistrationRepository;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.oauth2.client.authentication.OAuth2AuthenticationToken;
 import org.springframework.web.bind.annotation.*;
 
-import java.net.URI;
-import java.util.Collection;
-import java.util.HashMap;
-import java.util.Map;
 import java.util.Set;
 
 @RestController
@@ -26,46 +19,51 @@ import java.util.Set;
 @RequestMapping("/user")
 public class UserController {
 
-//    private JwtTokenUtil jwtTokenUtil;
-//    private JwtUserDetailsService userDetailsService;
-    private UserService userService;
-
     @Autowired
-    public UserController(
-//            JwtTokenUtil jwtTokenUtil,
-//                          JwtUserDetailsService userDetailsService,
-                          UserService userService) {
-//        this.jwtTokenUtil = jwtTokenUtil;
-//        this.userDetailsService = userDetailsService;
-        this.userService = userService;
-    }
-
-//    @PostMapping("/authenticate")
-//    public ResponseEntity<?> createAuthenticationToken(@RequestBody UserDTO userDTO) throws WrongValidationException {
-////        Collection<? extends GrantedAuthority> grantedAuthorities = userDetailsService.authenticateUserAndReturnAuthorities(userDTO);
-////        final String token = jwtTokenUtil.generateToken(userDTO.getUsername());
-////        User authenticationUser = userDetailsService.getAuthenticationUser(userDTO.getUsername());
-////        return ResponseEntity.ok(new JwtResponse(token, grantedAuthorities, userDTO.getUsername(), authenticationUser.getId()));
-//        return null;
-//    }
+    private UserService userService;
 
     @GetMapping
     public ResponseEntity<Set<UserDTO>> getAllUsers() {
         return ResponseEntity.ok(userService.getAll());
     }
 
-    @PostMapping
-    public Long createUser(@RequestBody UserDTO userDTO) throws Exception {
-        return userService.create(userDTO);
+    @GetMapping(value = "/info")
+    public ResponseEntity<UserDTO> getUserInfo(OAuth2AuthenticationToken authentication) {
+        String email;
+        try {
+            email = authentication.getPrincipal().getAttribute("email");
+        } catch (Exception e) {
+            throw new EmailMissingException("\"No email provided from the Oauth2 authentication service provider\"");
+        }
+        User user = userService.getUserByEmail(email);
+        UserDTO userDTO = UserMapper.INSTANCE.userToUserDTO(user);
+
+        return ResponseEntity.ok(userDTO);
     }
 
     @PutMapping
-    public void editUser(@RequestBody UserDTO userDTO) throws Exception {
+    public void editUser(@RequestBody UserDTO userDTO, OAuth2AuthenticationToken authentication) throws Exception {
+        validateUserHasAdminRole(authentication);
         userService.edit(userDTO);
     }
 
     @DeleteMapping(value = "/{id}")
-    public void deleteUser(@PathVariable Long id) throws Exception {
+    public void deleteUser(@PathVariable Long id, OAuth2AuthenticationToken authentication) throws Exception {
+        validateUserHasAdminRole(authentication);
         userService.delete(id);
+    }
+
+    private void validateUserHasAdminRole(OAuth2AuthenticationToken authentication){
+        String email;
+        try {
+            email = authentication.getPrincipal().getAttribute("email");
+        } catch (Exception e) {
+            throw new EmailMissingException("\"No email provided from the Oauth2 authentication service provider\"");
+        }
+        User user = userService.getUserByEmail(email);
+
+        if(user.getRoles().stream().noneMatch(role -> role.equals("admin"))){
+            throw new ForbiddenException();
+        }
     }
 }
