@@ -12,9 +12,7 @@ import java.util.Set;
 public class UserDTO {
     private Long id;
 
-    private String username;
-
-    private String password;
+    private String email;
 
     private List<String> roles = new ArrayList<>();
 
