@@ -6,11 +6,11 @@ import com.consultant.model.entities.User;
 import com.consultant.model.entities.Vacation;
 import com.consultant.model.exception.EntityAlreadyExists;
 import com.consultant.model.exception.NoMatchException;
+import com.consultant.model.exception.NotFoundException;
 import com.consultant.model.mappers.UserMapper;
 import com.consultant.model.repositories.UserRepository;
 import com.consultant.model.services.BasicOperationsService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import java.util.LinkedHashSet;
@@ -25,12 +25,10 @@ public class UserService implements BasicOperationsService<UserDTO> {
 
     private ConsultantService consultantService;
 
-    private PasswordEncoder passwordEncoder;
 
     @Autowired
-    public UserService(UserRepository userRepository, PasswordEncoder passwordEncoder, ConsultantService consultantService) {
+    public UserService(UserRepository userRepository, ConsultantService consultantService) {
         this.userRepository = userRepository;
-        this.passwordEncoder = passwordEncoder;
         this.consultantService = consultantService;
     }
 
@@ -49,16 +47,10 @@ public class UserService implements BasicOperationsService<UserDTO> {
         return userDTOS;
     }
 
+    // TODO fix this
     @Override
     public Long create(UserDTO userDTO) throws EntityAlreadyExists {
-        Optional<User> existingUser = userRepository.findByUsername(userDTO.getUsername());
-        if (existingUser.isPresent()) {
-            throw new EntityAlreadyExists("User already exists");
-        }
-        final User user = UserMapper.INSTANCE.userDTOToUser(userDTO);
-        user.setPassword(passwordEncoder.encode(user.getPassword()));
-
-        return userRepository.saveAndFlush(user).getId();
+        return null;
     }
 
     @Override
@@ -84,32 +76,22 @@ public class UserService implements BasicOperationsService<UserDTO> {
     }
 
     public User getUserByVacationId(Long vacationId) throws NoMatchException {
-        Optional<User> existingUser = userRepository.findByVacationId(vacationId);
-
-        return getUser(existingUser);
+        return userRepository.findByVacationId(vacationId)
+                .orElseThrow(() -> new NoMatchException("The id provided doesn't match any user"));
     }
 
     private void updateUser(User existingUser, UserDTO userDTO) {
-        if (userDTO.getPassword() != null) {
-            existingUser.setPassword(userDTO.getPassword());
-        }
         existingUser.setRoles(userDTO.getRoles());
-        existingUser.setUsername(userDTO.getUsername());
 
         userRepository.saveAndFlush(existingUser);
     }
 
     private User getExistingUserById(Long id) throws NoMatchException {
-        Optional<User> existingUser = userRepository.findById(id);
-
-        return getUser(existingUser);
+        return userRepository.findById(id)
+                .orElseThrow(() -> new NoMatchException("The id provided doesn't match any user"));
     }
 
-    private User getUser(Optional<User> existingUser) throws NoMatchException {
-        if (!existingUser.isPresent()) {
-            throw new NoMatchException("The id provided doesn't match any user");
-        }
-
-        return existingUser.get();
+    public User getUserByEmail(String email) {
+        return userRepository.findByEmail(email).orElseThrow(NotFoundException::new);
     }
 }
