@@ -15,9 +15,11 @@ import org.junit.runner.RunWith;
 import org.mockito.Mock;
 import org.mockito.Mockito;
 import org.mockito.junit.MockitoJUnitRunner;
-import org.springframework.security.crypto.password.PasswordEncoder;
 
-import java.util.*;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Optional;
+import java.util.Set;
 
 import static org.hamcrest.CoreMatchers.is;
 import static org.mockito.Mockito.*;
@@ -35,6 +37,8 @@ public class UserServiceShould {
 
     private Long userId = 1L;
 
+    private String userEmail = "user@mirado.com";
+
     private User user1 = new User();
 
     private User user2 = new User();
@@ -45,16 +49,13 @@ public class UserServiceShould {
 
     private Vacation vacation = new Vacation();
 
-    @Mock
-    private PasswordEncoder passwordEncoder;
-
     @Before
     public void setUp() {
-        userService = new UserService(userRepository,passwordEncoder,consultantService);
+        userService = new UserService(userRepository, consultantService);
 
-        Mockito.when(passwordEncoder.encode(Mockito.any())).thenReturn("123");
         user1 = new User();
         user1.setId(userId);
+        user1.setEmail(userEmail);
         when(userRepository.findAll()).thenReturn(userList);
 
         userDTO = new UserDTO();
@@ -80,29 +81,8 @@ public class UserServiceShould {
     }
 
     @Test
-    public void saveToRepositoryWhenCreatingUser() {
-        userDTO = new UserDTO();
-        userDTO.setId(userId);
-        Mockito.when(userRepository.saveAndFlush(Mockito.any())).thenReturn(user1);
-
-        userService.create(userDTO);
-
-        verify(userRepository, times(1)).saveAndFlush(Mockito.any());
-    }
-
-    @Test(expected = EntityAlreadyExists.class)
-    public void throwExceptionWhenCreatingUserWithSameUsername() throws EntityAlreadyExists {
-        String username = "username";
-        user1.setUsername(username);
-        userDTO.setUsername(username);
-        Mockito.when(userRepository.findByUsername(username)).thenReturn(Optional.ofNullable(user1));
-
-        userService.create(userDTO);
-    }
-
-    @Test
     public void deleteInRepositoryWhenDeletingExistingUser() throws NoMatchException {
-        Mockito.when(userRepository.findById(userId)).thenReturn(Optional.ofNullable(user1));
+        when(userRepository.findById(userId)).thenReturn(Optional.ofNullable(user1));
 
         userService.delete(userId);
 
@@ -111,16 +91,19 @@ public class UserServiceShould {
 
     @Test(expected = NoMatchException.class)
     public void throwExceptionWhenDeletingNonExistingUser() throws NoMatchException {
-        Mockito.when(userRepository.findById(userId)).thenReturn(Optional.empty());
+        when(userRepository.findById(userId)).thenReturn(Optional.empty());
 
         userService.delete(userId);
     }
 
     @Test
     public void updateOnEditingExistingUser() throws NoMatchException {
-        userDTO.setPassword("updated");
+        List<String> roles = new ArrayList<>();
+        roles.add("admin");
+
+        userDTO.setRoles(roles);
         userDTO.setId(userId);
-        Mockito.when(userRepository.findById(userId)).thenReturn(Optional.ofNullable(user1));
+        when(userRepository.findById(userId)).thenReturn(Optional.ofNullable(user1));
 
         userService.edit(userDTO);
 
@@ -129,9 +112,12 @@ public class UserServiceShould {
 
     @Test(expected = NoMatchException.class)
     public void throwExceptionWhenUpdatingNonExistingUser() throws NoMatchException {
-        userDTO.setPassword("updated");
+        List<String> roles = new ArrayList<>();
+        roles.add("admin");
+
+        userDTO.setRoles(roles);
         userDTO.setId(userId);
-        Mockito.when(userRepository.findById(userId)).thenReturn(Optional.empty());
+        when(userRepository.findById(userId)).thenReturn(Optional.empty());
 
         userService.edit(userDTO);
     }
@@ -139,7 +125,7 @@ public class UserServiceShould {
     @Test
     public void saveToRepositoryWhenUpdatingVacations() throws NoMatchException {
         vacation.setUserId(userId);
-        Mockito.when(userRepository.findById(userId)).thenReturn(Optional.ofNullable(user1));
+        when(userRepository.findById(userId)).thenReturn(Optional.ofNullable(user1));
 
         userService.updateUserVacations(vacation);
 
@@ -149,7 +135,7 @@ public class UserServiceShould {
     @Test(expected = NoMatchException.class)
     public void throwExceptionWhenUpdatingVacationsOfNonExistingUser() throws NoMatchException {
         vacation.setUserId(userId);
-        Mockito.when(userRepository.findById(userId)).thenReturn(Optional.empty());
+        when(userRepository.findById(userId)).thenReturn(Optional.empty());
 
         userService.updateUserVacations(vacation);
     }
@@ -159,11 +145,11 @@ public class UserServiceShould {
         Long vacationId = 1L;
         vacation.setUserId(userId);
         vacation.setId(vacationId);
-        Mockito.when(userRepository.findByVacationId(vacationId)).thenReturn(Optional.ofNullable(user1));
+        when(userRepository.findByVacationId(vacationId)).thenReturn(Optional.ofNullable(user1));
 
         User user = userService.getUserByVacationId(vacationId);
 
-        Assert.assertEquals(user1,user);
+        Assert.assertEquals(user1, user);
     }
 
     @Test(expected = NoMatchException.class)
@@ -171,7 +157,7 @@ public class UserServiceShould {
         Long vacationId = 1L;
         vacation.setUserId(userId);
         vacation.setId(vacationId);
-        Mockito.when(userRepository.findByVacationId(vacationId)).thenReturn(Optional.empty());
+        when(userRepository.findByVacationId(vacationId)).thenReturn(Optional.empty());
 
         userService.getUserByVacationId(vacationId);
     }
